ref(azure-devops): Migrate to commander subcommands
- Split monolithic index.ts into modular command files:
  - commands/configure.ts - org/project setup
  - commands/query.ts - run queries, change detection
  - commands/workitem.ts - fetch/display work items
  - commands/workitem-create.ts - creation wizard
  - commands/workitem-cache.ts - list cached items
  - commands/dashboard.ts - dashboard queries
  - commands/timelog.ts - placeholder for time logging

- Create shared cache.ts for DRY cache utilities:
  - Shared storage instance
  - CACHE_TTL constants per type
  - loadGlobalCache/saveGlobalCache functions
  - formatJSON utility

- Add QueryItemMetadata to types.ts
- Add --verbose flag for debug logging
- Use @app/azure-devops/* import aliases
