+++ src/azure-devops/commands/index.ts
@@ -1,0 +1,10 @@
+// Commands barrel file - exports all command registration functions
+// Each command module exports a function that registers its commands on the program
+
+export { registerConfigureCommand } from './configure';
+export { registerQueryCommand } from './query';
+export { registerWorkitemCommand } from './workitem';
+export { registerWorkitemCreateCommand } from './workitem-create';
+export { registerWorkitemCacheCommand } from './workitem-cache';
+export { registerDashboardCommand } from './dashboard';
+export { registerTimelogCommand } from './timelog';
